# swagger-output/swagger-specification.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "School Management System API",
    "version": "1.0.0",
    "description": "API for School Management System including authentication, academic year, subjects, exams, and mobile integrations",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "contact": {
      "name": "API Support",
      "url": "https://www.schoolmanagementsystem.com/support",
      "email": "support@schoolmanagementsystem.com"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "Term": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Term ID",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Term name",
            "example": "First Term"
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "description": "Term start date",
            "example": "2023-09-01T00:00:00.000Z"
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "description": "Term end date",
            "example": "2023-12-20T00:00:00.000Z"
          },
          "academicYearId": {
            "type": "integer",
            "description": "ID of the academic year this term belongs to",
            "example": 1
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Date and time when the term was created",
            "example": "2023-01-01T12:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Date and time when the term was last updated",
            "example": "2023-01-01T12:00:00.000Z"
          }
        }
      },
      "AcademicYear": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Academic year ID",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Academic year name",
            "example": "2023-2024"
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "description": "Academic year start date",
            "example": "2023-09-01T00:00:00.000Z"
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "description": "Academic year end date",
            "example": "2024-06-30T00:00:00.000Z"
          },
          "isDefault": {
            "type": "boolean",
            "description": "Whether this is the default academic year",
            "example": true
          },
          "terms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Term"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Date and time when the academic year was created",
            "example": "2023-01-01T12:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Date and time when the academic year was last updated",
            "example": "2023-01-01T12:00:00.000Z"
          }
        }
      },
      "CreateAcademicYearRequest": {
        "type": "object",
        "required": [
          "name",
          "startDate",
          "endDate"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Academic year name",
            "example": "2024-2025"
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "description": "Academic year start date",
            "example": "2024-09-01"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "description": "Academic year end date",
            "example": "2025-06-30"
          },
          "isDefault": {
            "type": "boolean",
            "description": "Whether this should be the default academic year",
            "example": false
          }
        }
      },
      "UpdateAcademicYearRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Academic year name",
            "example": "2024-2025 Updated"
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "description": "Academic year start date",
            "example": "2024-08-15"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "description": "Academic year end date",
            "example": "2025-06-15"
          },
          "isDefault": {
            "type": "boolean",
            "description": "Whether this should be the default academic year",
            "example": false
          }
        }
      },
      "TermRequest": {
        "type": "object",
        "required": [
          "name",
          "startDate",
          "endDate"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Term name",
            "example": "Second Term"
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "description": "Term start date",
            "example": "2024-01-10"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "description": "Term end date",
            "example": "2024-04-15"
          }
        }
      },
...
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/academic-years": {
      "get": {
        "summary": "Get all academic years",
        "description": "Retrieves a list of all academic years",
        "tags": [
          "Academic Years"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Number of items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "List of academic years retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "academicYears": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AcademicYear"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "integer"
                        },
                        "pages": {
                          "type": "integer"
                        },
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User is not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new academic year",
        "description": "Creates a new academic year with the provided details",
        "tags": [
          "Academic Years"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAcademicYearRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Academic year created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademicYear"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User is not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User does not have required permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/academic-years/{id}": {
      "get": {
        "summary": "Get academic year details",
        "description": "Retrieves details of a specific academic year by ID",
        "tags": [
          "Academic Years"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Academic year ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Academic year details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademicYear"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User is not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Academic year not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update an academic year",
        "description": "Updates an existing academic year with the provided details",
        "tags": [
          "Academic Years"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Academic year ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAcademicYearRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Academic year updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademicYear"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User is not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Academic year not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete an academic year",
        "description": "Deletes a specific academic year by ID",
        "tags": [
          "Academic Years"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Academic year ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Academic year deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User is not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Academic year not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/academic-years/{id}/default": {
      "post": {
        "summary": "Set an academic year as default",
        "description": "Sets a specific academic year as the default",
        "tags": [
          "Academic Years"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Academic year ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Academic year set as default successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "academicYear": {
                      "$ref": "#/components/schemas/AcademicYear"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User is not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User does not have required permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Academic year not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/academic-years/{id}/terms": {
      "post": {
        "summary": "Add a term to an academic year",
        "description": "Adds a new term to a specific academic year",
        "tags": [
          "Academic Years"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Academic year ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TermRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Term added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Term"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User is not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - User does not have required permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Academic year not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}